+++ inst/REFERENCES.bib
@@ -2,7 +2,7 @@
 address = {Boca Raton},
 author = {Hastie, Trevor and Tibshirani, Robert and Wainwright, Martin},
 booktitle = {CRC press},
-doi = {10.1201/b18401-1},
+doi = {10.1201/b18401},
 isbn = {978-1-4987-1217-0},
 issn = {0306-7734},
 mendeley-groups = {non{\_}reg},
@@ -362,8 +362,7 @@ year = {2018}
   number={1},
   pages={265--283},
   year={2007},
-  publisher={Institute of Mathematical Statistics},
-  doi = {10.1214/07-AOAS107}
+  publisher={Institute of Mathematical Statistics}
 }
 
 
@@ -479,8 +478,7 @@ year = {1996}
   number={428},
   pages={1535--1546},
   year={1994},
-  publisher={Taylor \& Francis Group},
-  doi = {10.21236/ada241408}
+  publisher={Taylor \& Francis Group}
 }
 
 
