+++ NAMESPACE
@@ -23,7 +23,7 @@ importFrom(MASS,mvrnorm)
 importFrom(bestglm,bestglm)
 importFrom(corpcor,cor2pcor)
 importFrom(corpcor,pcor2cor)
-importFrom(doSNOW,registerDoSNOW)
+importFrom(doParallel,registerDoParallel)
 importFrom(foreach,`%dopar%`)
 importFrom(foreach,foreach)
 importFrom(ggplot2,aes)
@@ -64,5 +64,6 @@ importFrom(stats,runif)
 importFrom(stats,sd)
 importFrom(stats,step)
 importFrom(stats,var)
+importFrom(utils,flush.console)
 importFrom(utils,setTxtProgressBar)
 importFrom(utils,txtProgressBar)
