### 2021-04-05 #3148 @donaldRwilliams
@danielskatz

I want to address the number of line of code, as the package appears meager.

That said, the package includes many analytic solutions, that I (with colleagues) have introduced to the social sciences. For example, one novel method is used to compute the expected replication of edges in large networks. This is derived analytically (few lines), whereas, if pursuing a simulation based approach, this would take many (many) more lines of code (but not necessary).

Further, to date, our three peer-reviewed papers describing the methods in this package have been cited over 100 times.
