### 2021-04-05 #3148 @donaldRwilliams
@danielskatz

Thanks !

Also note there are many other functions (e.g., 10 or so) that I took out for publishing version `1.0.0`, but I can happily add them back in, if desired.
