+++ README.Rmd
@@ -18,7 +18,8 @@ knitr::opts_chunk$set(
 
 # GGMnonreg: Non-regularized Gaussian Graphical Models
 
- 
+[![CRAN Version](http://www.r-pkg.org/badges/version/GGMnonreg)](https://cran.r-project.org/package=GGMnonreg)
+[![Downloads](https://cranlogs.r-pkg.org/badges/GGMnonreg)](https://cran.r-project.org/package=GGMnonreg)
 [![CircleCI build status](https://circleci.com/gh/donaldRwilliams/GGMnonreg.svg?style=svg)](https://circleci.com/gh/donaldRwilliams/GGMnonreg)
 
 
