### 2021-04-08 191200c
Add CRAN installation note
