+++ README.md
@@ -52,6 +52,12 @@ The following are also included
 
 ## Installation
 
+To install the latest release version (2.0.0) from CRAN use
+
+``` r
+install.packages("GGMnonreg")    
+```
+
 You can install the development version from
 [GitHub](https://github.com/) with:
 
@@ -125,16 +131,16 @@ fit <- ggm_inference(Y, boot = TRUE,
 
 confint(fit)
 #>              2.5%        97.5%
-#>  [1,] -0.29195928 -0.216827517
-#>  [2,] -0.15038354 -0.075762276
-#>  [3,]  0.25474519  0.327626779
-#>  [4,] -0.04330225  0.027234040
-#>  [5,]  0.12268992  0.217123536
-#>  [6,]  0.13522230  0.198745093
-#>  [7,] -0.06657910  0.001209683
-#>  [8,]  0.11345366  0.184490265
-#>  [9,]  0.34889721  0.411659437
-#> [10,]  0.08237053  0.160074874
+#>  [1,] -0.29767331 -0.213161936
+#>  [2,] -0.14495762 -0.078757052
+#>  [3,]  0.25689212  0.331357554
+#>  [4,] -0.04647112  0.028173576
+#>  [5,]  0.12102776  0.203929183
+#>  [6,]  0.13649686  0.206390457
+#>  [7,] -0.07925096  0.005588186
+#>  [8,]  0.10031299  0.193512433
+#>  [9,]  0.35044479  0.419958754
+#> [10,]  0.09322360  0.161286689
 ```
 
 These can then be plotted with, say, **ggplot2** (left to the user).
@@ -160,7 +166,7 @@ fit
 #> 1  1.00
 #> 2  1.00
 #> 3  1.00
-#> 4  0.02
+#> 4  0.10
 #> 5  1.00
 #> 6  1.00
 #> 7  0.35
@@ -293,8 +299,8 @@ cbind.data.frame(
 #> 2       1.000    1.000
 #> 3       1.000    1.000
 #> 4       1.000    1.000
-#> 5       0.918    0.912
-#> 6       0.053    0.052
+#> 5       0.895    0.912
+#> 6       0.050    0.052
 #> 7       0.000    0.000
 #> 8       0.000    0.000
 #> 9       0.000    0.000
@@ -302,7 +308,7 @@ cbind.data.frame(
 
 # average replicability (simulation)
 mean(R / length(index))
-#> [1] 0.4479402
+#> [1] 0.4464103
 
 # average replicability (analytic)
 fit_enr$ave_pwr
