### 2021-04-23 #3148 @donaldRwilliams
@arfon Sure, I can put those back in. How shall we do this ? I am currently out of town, but can certainly add them back in very very soon.

I wrote many of the functions that were removed in my first year of graduate school (more than 3 years ago), and I am now approaching my last year, so they were not concise and clearly written. That is why I removed them.

I can withdraw submission, and then resubmit ?

Or ?
