### 2021-05-04 #3148 @donaldRwilliams
@arfon

I wanted to follow up on this. As I mentioned, I can happily with draw the submission and resubmit at another time.

That said, I do want to bring to your attention (because I really like the mission of this journal) that I am reviewing a software paper for JOSS that is fewer lines of code (https://github.com/openjournals/joss-reviews/issues/2944). And, importantly, it is not only fewer lines, but the main functionality is to compute a mere **summary** of what my package provides.

On the other hand, my package here does not just do one thing or summarize the results of another package, but actually includes many methods, as shown here.

![image](https://user-images.githubusercontent.com/19657194/116960554-9b24a600-ac55-11eb-8075-9b42c2dc8ba4.png)
