### 2021-05-04 #3148 @arfon
Sorry for the delayed responses here @donaldRwilliams.

> Our papers describing the methods have been cited well over 100 times, with one in a flagship journal for psychometrics.

Could you say a little more about this please? Specifically, are descriptions of the methods implemented in this software already published in another journal?
