### 2021-05-04 #3148 @donaldRwilliams
@arfon

Not sure what you are asking. Can you clarify ?

Are you asking if the work has essentially been published elsewhere ? The answer is no.
