### 2021-05-04 #3148 @donaldRwilliams
@arfon

"descriptions of the methods implemented in this software already published in another journal"

Are you asking if this work has essentially been published elsewhere with this ? The answer is no.

(but again it is not entirely clear to me what you mean)

E.g., wouldn't we hope that most (if not all) statistical methods making their way into software are peer-reviewed themselves ? In fact, isn't that often the case ?
