### 2021-05-05 #3148 @arfon
Apologies, I should have been a little more explicit in my question here.

> Our papers describing the methods have been cited well over 100 times, with one in a flagship journal for psychometrics.

I'm responding to this comment ☝️ that you made above about there being papers describing the methods already in the literature.

> E.g., wouldn't we hope that most (if not all) statistical methods making their way into software are peer-reviewed themselves ? Isn't that often the case ?

In theory yes! But we're also sensitive to the idea of double-publishing for the same scholarly work, i.e., a paper describing the method, and a second (likely JOSS) paper describing the implementation of the method. In most cases, the work required for each of these is substantial, and so each paper in its own right represents a substantial scholarly contribution and so it's reasonable that there might be two papers.

In this situation, where the software contribution is definitely on the smaller side of what we consider allowable in JOSS (see https://blog.joss.theoj.org/2020/07/minimum-publishable-unit), whether or not there are already publications related to this one is a factor in our decision making here.

Hopefully this helps? I'm afraid I still don't have clarity on how this submission relates to this work: _"Our papers describing the methods have been cited well over 100 times, with one in a flagship journal for psychometrics."_
