### 2021-05-05 #3148 @donaldRwilliams
I am sorry, but I have to be frank.

I am quite literally reviewing a JOSS work that is actually **fewer lines** of code and computes a **summary** of what my package would provide. I have even consulted with the other reviewer, as we are both in the field, and he agreed that this is quite frustrating given the scope of this package compared to that one

The package here does the following
![image](https://user-images.githubusercontent.com/19657194/117140817-add8d100-ad62-11eb-80dd-7c054671779c.png)

Note I understand you are busy, but clearly having a package with fewer lines of code that computes one summary compared to a package that does all this seems a bit of an awkward situation (seems to not make much sense).

Anyhow, I am in quantitative methods in psychology. I often develop statistical methods. For example, in one case, we "invented" an approach to test differences in large networks. The software implementation is in this package. The primary paper did not even discuss the software.

As I mentioned, I can happily withdraw the package and resubmit.

Thanks !
