### 2021-05-05 #3148 @donaldRwilliams
@arfon

Note also I really much like the aims of this journal and respect very much the work you all do. And this frustration on my part will in no way compromise my review of that paper (https://github.com/openjournals/joss-reviews/issues/2944).

If you are going to reject the paper, please just do so with the option to resubmit.

I can just add the other functions in, when I have time, and we can start over :-)
