### 2021-05-05 #3148 @arfon
Hi @donaldRwilliams - I'm sorry that this is frustrating, and I understand why you are inclined to compare this submission with #2944. At this point though, the size of submissions (lines of code etc.) not at the core of my outstanding concern here.

My concern at this stage is getting clarity from you on the [question I asked above](https://github.com/openjournals/joss-reviews/issues/3148#issuecomment-832520828) which I will ask again for clarity:

How does this submission relate to any other published work _about the same method and or software_?

In the [above comment](https://github.com/openjournals/joss-reviews/issues/3148#issuecomment-825612955) you say _"Our papers describing the methods have been cited well over 100 times, with one in a flagship journal for psychometrics."_. This comment raises a concern of potential duplicate publishing of the same work which would mean we could not review this work.
