### 2021-05-06 #3148 @donaldRwilliams
Hmm.

As I have said, there is no duplication.  This is still not entirely clear to me

"How does this submission relate to any other published work about the same method and or software?"

To clarify, I am happy to zoom with you.

The package estimates *6* different kinds of graphical models, and it is the only R package to do so. And there is a need to do so in my field. The methods for estimating graphical models go back decades (not our work), but there is no software that implements them all. Perhaps if we searched hard, we can find some implementations of some of them scattered about (honestly not sure, as my search came up empty). But, really, this is the only package that is one-stop shop for graphical modeling.

Then there are some additional methods, e.g., for testing differences in those graphical models, maybe quantifying expected replicability, etc. These are the ones we introduced to the field, but are not the focus of the package at all. The focus is graphical modeling, which is becoming very, very popular in my field.

Does this help ?
