### 2021-05-06 #3148 @donaldRwilliams
Note also that I mentioned my slight frustration to you in an effort to improve an already great journal. As I mentioned, I have consulted with others that know of my work, who are in the field, etc. and they felt that I should say something as it seemed a bit inconsistent.  But I think perhaps this is due to maybe a lack of clarity arising from not being sure what the other means, etc.
