### 2021-05-07 #3148 @arfon
> Perhaps from my previous comment you may have thought the package only implements what is in our papers ?

That's exactly what I was concerned about but I realize that's not the case now. Thank you for your patience explaining this @donaldRwilliams.
