### 2021-05-07 #3148 @donaldRwilliams
> > Perhaps from my previous comment you may have thought the package only implements what is in our papers ?
>
> That's exactly what I was concerned about but I realize that's not the case now. Thank you for your patience explaining this @donaldRwilliams.

Thank you as well for working through this with me. I realize now that I was not clear about what the package did, etc.

Anyhow, here is another reviewer.

@AlexChristensen

And also

@paytonjjones

If those do not work out, I can provide some others.
