+++ inst/REFERENCES.bib
@@ -876,3 +876,34 @@ year = {2011}
   year={2013},
   publisher={Wiley Online Library}
 }
+
+@article{marsman2018introduction,
+  title={An introduction to network psychometrics: Relating Ising network models to item response theory models},
+  author={Marsman, M and Borsboom, D and Kruis, J and Epskamp, S and Van Bork, R and Waldorp, LJ and Maas, HLJ van der and Maris, G},
+  journal={Multivariate behavioral research},
+  volume={53},
+  number={1},
+  pages={15--35},
+  year={2018},
+  publisher={Taylor \& Francis}
+}
+
+@article{lenz1920beitrvsge,
+  title={Beitr{\v{s}}ge zum verst{\v{s}}ndnis der magnetischen eigenschaften in festen k{\v{s}}rpern},
+  author={Lenz, Wilhelm},
+  journal={Physikalische Z},
+  volume={21},
+  pages={613--615},
+  year={1920}
+}
+
+@article{ising1925beitrag,
+  title={Beitrag zur theorie des ferromagnetismus},
+  author={Ising, Ernst},
+  journal={Zeitschrift f{\"u}r Physik},
+  volume={31},
+  number={1},
+  pages={253--258},
+  year={1925},
+  publisher={Springer}
+}
