### 2021-05-07 #3148 @chrisbrunsdon
Is it possible to remove me from this? I now have 13000 or so emails in mail box and only around 300 not openjournals/jossreviews related...

Sent from my iPhone

> On 7 May 2021, at 8:15 p.m., Payton J. Jones ***@***.***> wrote:
>
> ﻿
> Unfortunately no, I'm out of academia now and wouldn't be able to get a
> review done in a timely manner.
>
> On Fri, May 7, 2021 at 11:44 AM Arfon Smith ***@***.***>
> wrote:
>
> > 👋 @AlexChristensen <https://github.com/AlexChristensen> and @paytonjjones
> > <https://github.com/paytonjjones> - would you be willing to review this
> > submission for JOSS? We carry out our checklist-driven reviews here in
> > GitHub issues and follow these guidelines:
> > https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
> >
> > The submission under review here is *GGMnonreg: Non-Regularized Gaussian
> > Graphical Models* <https://github.com/donaldRwilliams/GGMnonreg>.
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <https://github.com/openjournals/joss-reviews/issues/3148#issuecomment-834686181>,
> > or unsubscribe
> > <https://github.com/notifications/unsubscribe-auth/AFQWBJKD7BRDQ64W2ZEGR7LTMQYI3ANCNFSM42M6QKEA>
> > .
> >
