### 2021-05-22 #3148 @arfon
:wave: @hfgolino @Jillderon – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [_GGMnonreg: Non-Regularized Gaussian Graphical Models_](https://github.com/donaldRwilliams/GGMnonreg).
