### 2021-05-22 #3148 @GiulioCostantini
Hi all,
Sorry, I missed the first email. I would like to review this package, when
is the deadline for completing this task?
Best,
Giulio

Giulio Costantini, PhD
Assistant professor
Psychology Department
University of Milano-Bicocca
P.zza dell'Ateneo Nuovo, 1 (U6) - room 3127
20126 Milan (MI)
Phone ++39 02.6448.3839

Il giorno ven 21 mag 2021 alle ore 19:47 Sacha Epskamp <
***@***.***> ha scritto:

> Hi! I would love to review this package as it looks like an excellent
> package indeed. Unfortunately, however, I have paternity leave coming up
> very fast now and too many deadlines before to finish. Some people that
> come to mind perhaps in addition to Giulio are Hudson Golino, Jonas
> Haslbeck, Alex Millner, Jill de Ron & Myrthe Veenman.
>
