### 2021-05-22 #3148 @arfon
> I would like to review this package

⚡ thanks!

> when is the deadline for completing this task?

@GiulioCostantini - ideally we'd want the review to be complete within 4-6 weeks but we ask reviewers to start before then as JOSS reviews are generally quite iterative (authors making updates as feedback is received).

I'm going to add you as a reviewer now and start the main review issue. If this timing doesn't work, please let me know, we can probably work with your schedule.
