### 2021-06-22 #3308 @AlexChristensen
Hi @donaldRwilliams,

Here's a couple comments on my pass through the code:

o The return values for the `returned_object` for several functions code be documented with greater precision. Take `enr` for example:

```
returned_object <- list(ave_pwr = ave_pwr,
                          cdf = cdf,
                          p_s = p_s,
                          p = p,
                          n_nonzero = n_nonzero,
                          n = n,
                          replications = replications,
                          var_pwr = var_pwr,
                          type = type)
```

Each one of these could be listed in the documentation:

```
#' @return An list of class \code{enr}, including a variety of information used
#' by other functions (e.g., to plot the results). These objects are:
#'
#' \itemize{
#'
#' \item{ave_pwr}{Average power}
#'
#' \item{cdf}{Cumulative distribution function of values greater than...}
#'
#' \item{p_s}{Probability of success}
#'
#' \item{n_nonzero}{Non-zero edges}
#'
#' ...
#'
#' \item{type}{Type of correlation coefficient}
#'
#' }
```

This could be done for most functions, which would help people better understand the output. Another option is to make the output's object names be exactly what they are. For example:

`avg_pwr` = `average_power`

`cdf` = `cumulative_distribution_function`

etc.

Finally, you could stick with what you have for some (e.g., `gen_net`) and just add the output in parentheses:

```
#' @return A list with the true structure (pcors), adjacency matrix (adj),
#' number of attempts (trys) and correlation matrix (cors).
```

o I like to have a package documentation page pop up for the name of the package. For example `?GGMnonreg`

This is mostly a convenience preference but when I don't know any of the names of the packages functions, I'll first try `?package_name` before scrolling through all my packages

Here's an example from our _EGAnet_ package: https://github.com/hfgolino/EGAnet/blob/master/R/EGAnet.R

I have not had a chance to look over the document but I will do so later today. I'll go through the checklist once I wrap up there
