### 2021-06-22 #3308 @AlexChristensen
@donaldRwilliams,

Here are some comments for the other parts of the checklist:

## Documentation (GitHub Readme)

o Latest CRAN version is 1.0.0 (says 2.0.0 in documentation)

## Paper

o I would move the third paragraph of the "Summary" section to the "Statement of Need." It provides the best reason why there is a need for these low-dimensional data methods.

o In your `plot` S3methods, consider moving `get_graph` into the S3method so people can just drop the object in `plot`. For example

```
plot(get_graph(fit), node_names = colnames(Y), edge_magnify = 2)
```

to work instead as:

```
plot(fit, node_names = colnames(Y), edge_magnify = 2)
```

For `plot.graph`, this could be executed as `x <- get_graph(x)`

o "State of the Field" section is blended in the "Summary" section. This is OK by me (checking this box on the checklist)

o The paper could be a bit more robust. Mostly just adding a bit more description about the data and including some of the examples that are included in the documentation: `predictability`, `confint`, `eip`, and `enr`. These are all functions people will want to use, so I would highlight them for maximum user-friendliness in the paper too.
