### 2021-06-25 #3308 @GiulioCostantini
@donaldRwilliams
First I have to say that I really appreciated your work. I agree with all comments made by Alex. Going through the help files, I found it crucial to include more documentation of specific items in the output of each function. The paper is very short, but I understand that this is part of JOSS strict word limit policy. I have a few suggestions and comments:
- Why did you include, under the asd_ocd dataset examples, the code for simluating from such data using MASS? This does not seem directly relevant for your package.
- Why is the default number of cores in ggm_inference equal to 2? My suggestion would be to make it 1, to work on any machine, or parallel::detectCores()-1, if it is important to enable mutlicore processing by default, unless there are important reasons to have 2 cores.
- Function constrained: I honestly could not understand what this function does. The referenced book is quite broad: perhaps a more specific reference (e.g., a page or paragraph number) would help.
- Function eip: If there are papers in which this was used, perhaps include the references in the .Rd file.
- Function enr: Perhaps more information in the Description or Details section of the .Rd file would be beneficial. Similarly, I found the example section not self-explanatory and a bit difficult to follow. perhaps some more comments would help.
- functon gen_net. The title in the .Rd file would perhaps be clearer if it was something like "simulate a partial correlation matrix" instead of "generate true...".
- In the help file of function ggm_compare, perhaps it would be important to specify what it is meant by "compare" (e.g., establish whether each of the corresponding edges are significantly different in two groups). Also, instead of including "method" as an argument, perhaps you could use the "..." syntax and allow passing arguments to ggm_inference, if this does not have problematic implications.

Best,
Giulio
