### 2021-06-25 #3308 @donaldRwilliams
@GiulioCostantini

Thank you so much for the great suggestions  !

I will make all of the changes you suggested  !

I am confident that will greatly improve the package  :-)
