### 2021-11-05 #3308 @donaldRwilliams
work in progress

- Why did you include, under the asd_ocd dataset examples, the code for simluating from such data using MASS? This does not seem directly relevant for your package.

It is a correlation matrix.

I see now the title is misleading (“Data”). I changed that to Cor.

- Why is the default number of cores in ggm_inference equal to 2? My suggestion would be to make it 1, to work on any machine, or parallel::detectCores()-1, if it is important to enable mutlicore processing by default, unless there are important reasons to have 2 cores.

Great catch !

I changed it to 1

- Function constrained: I honestly could not understand what this function does. The referenced book is quite broad: perhaps a more specific reference (e.g., a page or paragraph number) would help.

Fair enough.

I added a more lengthy description, and the exact page in the reference.
