+++ inst/REFERENCES.bib
@@ -1,3 +1,13 @@
+@article{emmert2019constrained,
+  title={Constrained covariance matrices with a biologically realistic structure: Comparison of methods for generating high-dimensional Gaussian graphical models},
+  author={Emmert-Streib, Frank and Tripathi, Shailesh and Dehmer, Matthias},
+  journal={Frontiers in Applied Mathematics and Statistics},
+  volume={5},
+  pages={17},
+  year={2019},
+  publisher={Frontiers}
+}
+
 @book{Hastie2015,
 address = {Boca Raton},
 author = {Hastie, Trevor and Tibshirani, Robert and Wainwright, Martin},
