+++ DESCRIPTION
@@ -3,7 +3,7 @@ Type: Package
 Title: Non-Regularized Gaussian Graphical Models
 Version: 1.0.1
 Date: 2021-05-07
-Authors@R:  person("Donald", "Williams", email = "drwwilliams@ucdavis.edu", role = c("aut", "cre"))
+Authors@R: person("Donald", "Williams", email = "drwwilliams@ucdavis.edu", role = c("aut", "cre"))
 Maintainer: Donald Williams <drwwilliams@ucdavis.edu>
 Description: Estimate non-regularized Gaussian graphical models, Ising models, 
   and mixed graphical models. The current methods consist of multiple regression, 
@@ -35,4 +35,4 @@ Suggests:
  qgraph
 LazyData: true
 RdMacros: Rdpack
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.2
