### 2021-11-06 3ddbc18
final two comments for JOSS (R1)
