+++ man/enr.Rd
@@ -26,17 +26,25 @@ An list of class \code{enr}, including a variety of information used
         by other functions (e.g., to plot the results).
 }
 \description{
-Compute expected network replicability in any number of
-             replication attempts. This works for any kind of
-             correlation, assuming it is possible to obtain the
-             standard error (analytically or with a bootstrap).
+Investigate network replicability for any kind of
+             partial correlation, assuming there is an analytic
+             solution for the standard error (e.g., Pearson's or Spearman's).
 }
 \note{
-This method was introduced in \insertCite{williams2020learning;textual}{GGMnonreg}.
+This method was introduced in
+\insertCite{williams2020learning;textual}{GGMnonreg}.
+
+The basic idea is to determine the replicability of edges in a
+partial correlation network. This requires defining the true
+network, which can include edges of various sizes, and then
+solving for the proportion of edges that are expected
+to be replicated (e.g. in two, three, or four replication attempt).
 }
 \examples{
 \donttest{
-# correlations
+# (1) define partial correlation network
+
+# correlations from ptsd symptoms
 cors <- cor(GGMnonreg::ptsd)
 
 # inverse
@@ -46,24 +54,35 @@ inv <- solve(cors)
 pcors <-  -cov2cor(inv)
 
 # set values to zero
+# (this is the partial correlation network)
 pcors <- ifelse(abs(pcors) < 0.05, 0, pcors)
 
-fit_enr <- enr(net = pcors, n = 500, replications = 2)
+
+# compute ENR in two replication attempts
+fit_enr <- enr(net = pcors,
+               n = 500,
+               replications = 2)
 
 
-# intuition for the method
+# intuition for the method:
+# The above did not require simulation, and here I use simulation
+# for the same purpose.
 
 # location of edges
+# (where the edges are located in the network)
 index <- which(pcors[upper.tri(diag(20))] != 0)
 
-# convert network into correlation matrix
+# convert network a into correlation matrix
+# (this is needed to simulate data)
 diag(pcors) <- 1
 cors_new <- corpcor::pcor2cor(pcors)
 
 # replicated edges
+# (store the number of edges that were replicated)
 R <- NA
 
-# increase 100 to, say, 5,000
+# simulate how many edges replicate in two attempts
+# (increase 100 to, say, 5,000)
 for(i in 1:100){
 
   # two replications
@@ -93,12 +112,12 @@ cbind.data.frame(
   data.frame(analytic = round(fit_enr$cdf, 3))
 )
 
+# now compare simulation to the analytic solution
 # average replicability (simulation)
 mean(R / length(index))
 
 # average replicability (analytic)
 fit_enr$ave_pwr
-
 }
 
 }
