### 2021-11-07 #3308 @donaldRwilliams
**Reviewer 2**

- The return values for the returned_object for several functions code be documented with greater precision. Take enr for example

Agreed !

I went through each function and did just as you suggested
