+++ R/mixed_search.R
@@ -11,8 +11,24 @@
 #'           \code{"AIC"} and \code{"BIC"} (default).
 #'
 #'
-#' @return An object of class \code{mixed_search}, including \code{wadj} (weighted adjacency matrix)
-#'         and \code{adj} (adjacency matrix).
+#' @return An object of class \code{mixed_search} including
+#'
+#' \itemize{
+#'
+#' \item{\strong{wadj}}: Weighted adjacency matrix, corresponding
+#' to the partial correlation network.
+#'
+#' \item{\strong{adj}}: Adjacency matrix (detected effects).
+#'
+#' \item{\strong{pcors}}: Partial correlations.
+#'
+#' \item{\strong{n}}: Sample size.
+#'
+#' \item{\strong{p}}: Number of nodes.
+#'
+#' \item{\strong{Y}}: Data.
+#'
+#' }
 #'
 #' @details Only backwards selection is currently implemented.
 #'          Only an adjacency matrix is provided.
