+++ man/plot_enr.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/enr.R
-\name{plot.enr}
-\alias{plot.enr}
+\name{plot_enr}
+\alias{plot_enr}
 \title{Plot \code{enr} Objects}
 \usage{
-\method{plot}{enr}(x, iter = 1e+05, fill = "#009E73", alpha = 0.5, ...)
+plot_enr(x, iter = 1e+05, fill = "#009E73", alpha = 0.5, ...)
 }
 \arguments{
 \item{x}{An object of class \code{enr}.}
@@ -42,7 +42,7 @@ pcors <- ifelse(abs(pcors) < 0.05, 0, pcors )
 est <- enr(net = pcors, n = 500, replications = 2)
 
 # plot
-plot(est)
+plot_enr(est)
 }
 
 }
