+++ man/plot.ggmnonreg.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/plot_graph.R
-\name{plot.graph}
-\alias{plot.graph}
+\name{plot.ggmnonreg}
+\alias{plot.ggmnonreg}
 \title{Network Plot for \code{graph} Objects}
 \usage{
-\method{plot}{graph}(
+\method{plot}{ggmnonreg}(
   x,
   layout = "circle",
   neg_col = "#D55E00",
@@ -58,6 +58,6 @@ Y <- ptsd
 # estimate graph
 fit <- ggm_inference(Y, boot = FALSE)
 
-# get info for plotting
-plot(get_graph(fit))
+# plot graph
+plot(fit)
 }
