+++ inst/REFERENCES.bib
@@ -917,3 +917,12 @@ year = {2011}
   year={1925},
   publisher={Springer}
 }
+
+@article{williams_2021_conf,
+ title={The Confidence Interval that Wasn’t: Bootstrapped “Confidence Intervals” in L1-Regularized Partial Correlation Networks},
+ url={psyarxiv.com/kjh2f},
+ DOI={10.31234/osf.io/kjh2f},
+ journal={PsyArXiv},
+ author={Williams, Donald R},
+ year={2021}
+}
