+++ NAMESPACE
@@ -1,8 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(confint,ggm_inference)
-S3method(plot,enr)
-S3method(plot,graph)
+S3method(plot,ggmnonreg)
 S3method(print,ggmnonreg)
 export(constrained)
 export(eip)
@@ -16,6 +15,7 @@ export(ggm_inference)
 export(ggm_search)
 export(ising_search)
 export(mixed_search)
+export(plot_enr)
 export(predictability)
 importFrom(GGMncv,constrained)
 importFrom(GGMncv,ggmncv)

[…]
