+++ paper.md
@@ -7,12 +7,13 @@ tags:
 - Ising model
 authors:
   - name: Donald R. Williams
-    affiliation: 1 # (Multiple affiliations must be quoted)
+    affiliation: "1", "2"
 affiliations:
  - name: Department of Psychology, University of California, Davis
    index: 1
+ - name: NWEA, Portland, USA  
 citation_author: Williams
-date: 05 April 2021
+date: 08 November 2021
 year: 2021
 bibliography: inst/REFERENCES.bib
 ---
@@ -24,14 +25,12 @@ Studying complex relations in multivariate datasets is a common task across the 
 Graphical modeling is quite common in fields with wide data, that is, when there are more variables ($p$) than observations ($n$). Accordingly, many regularization-based approaches have been developed for those kinds of data. There are key drawbacks of regularization, including, but not limited to, the fact that obtaining a valid measure of parameter uncertainty is very (very) difficult [@Buhlmann2014] and there can
 be an inflated false positive rate [see for example, @williams2019nonregularized].
 
-More recently, graphical modeling has emerged in psychology (Epskamp et al. 2018), where the data is typically long or low-dimensional ($p < n$; @williams2019nonregularized, @williams_rethinking). The primary purpose of GGMnonreg is to provide methods that were specifically designed for low-dimensional data (e.g., those common in the social-behavioral sciences).
-
-
 # Statement of Need
-The following were designed specifically for low-dimensional data, for which there is a dearth of methodology.
+More recently, graphical modeling has emerged in psychology (Epskamp et al. 2018), where the data is typically long or low-dimensional ($p < n$; @williams2019nonregularized, @williams_rethinking). The primary purpose of 
+**GGMnonreg** is to provide methods that were specifically designed for low-dimensional data (e.g., those common in the social-behavioral sciences).
 
 ## Supported Models
-* Gaussian graphical model. The following data types are supported.
+* Gaussian graphical model (GGM). The following data types are supported.
   + Gaussian 
   + Ordinal
   + Binary
@@ -49,7 +48,44 @@ The following are also included
 * Constrained precision matrix [the network, given an assumed graph, see p. 631 in @hastie2009elements]
 * Predictability [variance explained for each node, @Haslbeck2018]
 
-## Implementation
+## Gaussian graphical Model
+The following estimates a GGM for 5 post-traumatic stress disorder 
+(PTSD) symptoms [@armour2017network]:
+
+```{r}
+fit <- ggm_inference(Y = ptsd[,1:5], 
+                     boot = FALSE)
+
+fit
+#>           1         2         3         4         5
+#> 1 0.0000000 0.2262934 0.0000000 0.3335737 0.1547986
+#> 2 0.2262934 0.0000000 0.4993419 0.0000000 0.0000000
+#> 3 0.0000000 0.4993419 0.0000000 0.2205442 0.1841798
+#> 4 0.3335737 0.0000000 0.2205442 0.0000000 0.3407634
+#> 5 0.1547986 0.0000000 0.1841798 0.3407634 0.0000000
+```
+
+### Predictability
+It is common to then estimate "predictability", which corresponds to $R^2$
+for each node in the network. In **GGMnonreg**, this is implemented with the 
+following code:
+
+```{r}
+predictability(fit)
+
+#>   Estimate Est.Error Ci.lb Ci.ub
+#> 1     0.45      0.05  0.35  0.54
+#> 2     0.50      0.05  0.41  0.59
+#> 3     0.55      0.04  0.47  0.64
+#> 4     0.50      0.05  0.41  0.59
+#> 5     0.46      0.05  0.37  0.55
+```
+
+
+## Ising Model
+An Ising model is for binary data. The PTSD symptoms can be binary, indicating
[…]
