+++ paper.md
@@ -7,10 +7,12 @@ tags:
 - Ising model
 authors:
   - name: Donald R. Williams
-    affiliation: 1
+    affiliation: "1, 2"
 affiliations:
  - name: Department of Psychology, University of California, Davis
    index: 1
+ - name: NWEA, Portland, USA
+   index:
 citation_author: Williams
 date: 08 November 2021
 year: 2021
