### 2021-11-08 #3308 @donaldRwilliams
@GiulioCostantini @AlexChristensen @arfon

I want to thank you all for baring with me (google internship). I put a pause on basically all papers in review, etc. and actually withdrew all but this one and another at the R journal (because I find software the most useful thing I work on !)

Anyhow, I have address all the comments
