+++ R/eip.R
@@ -42,15 +42,20 @@ eip <- function(Y,
                 method = "pearson",
                 B = 1000,
                 progress = TRUE) {
+
   n <- nrow(Y)
+
   p <- ncol(Y)
 
   if (progress) {
-    pb <- utils::txtProgressBar(max = B, style = 3)
+    pb <- utils::txtProgressBar(max = B,
+                                style = 3)
   }
+
   eips <- t(sapply(1:B, function(x) {
     eips_i <-  ggm_inference(Y[sample(1:n, n, replace = T), ],
-                             boot = FALSE, method = method)$adj[upper.tri(diag(p))]
+                             boot = FALSE,
+                             method = method)$adj[upper.tri(diag(p))]
 
     if (progress) {
       setTxtProgressBar(pb, x)
@@ -60,7 +65,9 @@ eip <- function(Y,
   }))
 
   returned_object <- list(eip =  data.frame(eip = colMeans(eips)))
-  class(returned_object) <- c("ggmnonreg", "eip")
+
+  class(returned_object) <- c("ggmnonreg",
+                              "eip")
   return(returned_object)
 }
 
