+++ R/mixed_search.R
@@ -69,13 +69,13 @@ mixed_search <- function(Y, data_type = NULL, IC = "BIC"){
   }
 
   estimates <- lapply(1:p, function(x) {
-    dat <- cbind.data.frame(X[,-x], y = X[, x])
+    dat <- cbind.data.frame(X[, -x], y = X[, x])
 
-    if(data_type[x] == "b"){
+    if (data_type[x] == "b") {
       fit <- glm(y ~ ., data = dat, family = binomial())
-    } else if (data_type[x] == "p"){
+    } else if (data_type[x] == "p") {
       fit <- glm(y ~ ., data = dat, family = poisson())
-    } else if(data_type[x] == "g"){
+    } else if (data_type[x] == "g") {
       fit <- glm(y ~ ., data = dat, family = gaussian())
     } else {
       stop("'data_type' must be 'b', 'p' or 'g'")
