+++ inst/REFERENCES.bib
@@ -927,14 +927,3 @@ year = {2011}
  year={2021}
 }
 
-
-@article{armour2017network,
-  title={A network analysis of DSM-5 posttraumatic stress disorder symptoms and correlates in US military veterans},
-  author={Armour, Cherie and Fried, Eiko I and Deserno, Marie K and Tsai, Jack and Pietrzak, Robert H},
-  journal={Journal of anxiety disorders},
-  volume={45},
-  pages={49--59},
-  year={2017},
-  publisher={Elsevier},
-  doi={10.1016/j.janxdis.2016.11.008}
-}

[…]
