### 2021-11-09 18d6c8a
document gen_pcors
