### 2021-11-09 d8a9b00
paper: remove get_graph
