### 2021-11-10 a6f5776
add zenodo badge and add new hex
