+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-<img src="man/figures/nonreg_hex.png" width = 200 />
+<img src="man/figures/nonreg_hex.png" width = 250 />
 
 # GGMnonreg: Non-regularized Gaussian Graphical Models
 
@@ -132,16 +132,16 @@ fit <- ggm_inference(Y, boot = TRUE,
 
 confint(fit)
 #>              2.5%        97.5%
-#>  [1,] -0.28812453 -0.225070244
-#>  [2,] -0.15246849 -0.069719684
-#>  [3,]  0.24575339  0.329247067
-#>  [4,] -0.04745866  0.031420127
-#>  [5,]  0.12302600  0.195448033
-#>  [6,]  0.12215106  0.215509020
-#>  [7,] -0.06441601  0.009372784
-#>  [8,]  0.10747029  0.182046649
-#>  [9,]  0.33526851  0.413057325
-#> [10,]  0.06704818  0.160007174
+#>  [1,] -0.29205233 -0.212951642
+#>  [2,] -0.14926375 -0.075919715
+#>  [3,]  0.25508466  0.326055944
+#>  [4,] -0.03778214  0.032287534
+#>  [5,]  0.12219960  0.204592760
+#>  [6,]  0.13387070  0.206412243
+#>  [7,] -0.07641808  0.006226351
+#>  [8,]  0.10997005  0.195192621
+#>  [9,]  0.34388225  0.418927425
+#> [10,]  0.08130533  0.153621543
 ```
 
 These can then be plotted with, say, **ggplot2** (left to the user).
@@ -170,7 +170,7 @@ fit
 #> 4  0.06
 #> 5  1.00
 #> 6  1.00
-#> 7  0.29
+#> 7  0.31
 #> 8  1.00
 #> 9  1.00
 #> 10 1.00
@@ -300,8 +300,8 @@ cbind.data.frame(
 #> 2       1.000    1.000
 #> 3       1.000    1.000
 #> 4       1.000    1.000
-#> 5       0.903    0.912
-#> 6       0.054    0.052
+#> 5       0.897    0.912
+#> 6       0.055    0.052
 #> 7       0.000    0.000
 #> 8       0.000    0.000
 #> 9       0.000    0.000
@@ -309,7 +309,7 @@ cbind.data.frame(
 
 # average replicability (simulation)
 mean(R / length(index))
-#> [1] 0.4475299
+#> [1] 0.4482051
 
 # average replicability (analytic)
 fit_enr$ave_pwr
